Seperated classes into individual files and created a run script.
